requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.5.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
paho-mqtt==2.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
